#38 chore: Refactor ingestion command code
